Extract images/networks/vms mixins from BoxmanManager (#85 item 29, part 2)
Move-only refactor: infrastructure method groups move verbatim into
mixin classes under src/boxman/manager_parts/ (ImagesMixin,
NetworksMixin, VMsMixin); the module-level multiprocessing target
_clone_with_retry moves with the vms group. The only in-code adjustment
is _clone_with_retry's reference to the _is_diskless_boot staticmethod,
now resolved via ImagesMixin instead of BoxmanManager (same function
object — boxman.manager cannot be imported from the mixin modules
without a circular import).

Tests: patch anchors that targeted names rehomed to the new modules were
repointed (boxman.manager.VirshCommand/ImageCache/Process/run ->
boxman.manager_parts.{images,vms}.*). No test logic changed.

Refs #85
