+++ src/boxman/manager.py
@@ -1,14 +1,10 @@
-import contextlib
-import hashlib
 import json
-import logging
 import os
 import shutil
 import subprocess
 import time
 from multiprocessing import Process, Queue
 from typing import Any, Optional
-from urllib.parse import urlparse
 
 import yaml
 
@@ -16,22 +12,19 @@ from boxman import log
 from boxman.abstract.providers import ProviderSession
 from boxman.config_cache import BoxmanCache
 from boxman.exceptions import ConfigError, ProvisionError, SnapshotError
-from boxman.image_cache import ImageCache
-from boxman.loggers.logger import suppressed
 from boxman.manager_parts.config import ConfigMixin
+from boxman.manager_parts.images import ImagesMixin
 from boxman.manager_parts.naming import NamingMixin
+from boxman.manager_parts.networks import NetworksMixin
 from boxman.manager_parts.ssh import SSHMixin
+from boxman.manager_parts.vms import VMsMixin
 from boxman.manager_parts.workspace import WorkspaceMixin
 from boxman.netlab import ContainerlabManager, shared_bridges
 from boxman.providers import merge_provider_configs, primary_provider_type
-from boxman.providers.libvirt import net_reconcile
 from boxman.providers.libvirt.commands import VirshCommand
-from boxman.providers.libvirt.virsh_parse import parse_domblklist
 from boxman.runtime import RuntimeBase, create_runtime
 from boxman.task_runner import TaskRunner
-from boxman.utils.http_download import download_url
 from boxman.utils.jinja_env import create_jinja_env
-from boxman.utils.shell import run
 
 #: Seconds to wait for a finished child's queue message before declaring its
 #: result missing (the child has already joined, so any wait here is just
@@ -53,52 +46,9 @@ def _parallel_worker(result_queue, label, target, args):
         result_queue.put((label, False, f"{type(exc).__name__}: {exc}"))
 
 
-def _clone_with_retry(provider, cluster, vm_info, new_vm_name,
-                      max_retries: int = 5) -> None:
-    """
-    Clone one VM, retrying transient (e.g. storage-pool-busy) failures.
-
-    Module-level — not a method or closure — so it stays picklable as a
-    ``multiprocessing.Process`` target. Shared by
-    :meth:`BoxmanManager.clone_vms` and
-    :meth:`BoxmanManager._clone_and_configure_new_vms`.
-    """
-    for attempt in range(1, max_retries + 1):
-        last_attempt = attempt == max_retries
-        # Suppress error-level logs on all retryable attempts so that
-        # transient pool-busy failures don't appear as errors; only the
-        # final attempt logs errors normally. suppressed() restores the
-        # prior level (not a hardcoded DEBUG) so -v/-vv survives retries.
-        _cm = (contextlib.nullcontext() if last_attempt
-               else suppressed(logging.CRITICAL))
-        try:
-            with _cm:
-                src_vm_name = vm_info.get('base_image') or cluster.get('base_image')
-                if not src_vm_name and not BoxmanManager._is_diskless_boot(vm_info):
-                    raise ValueError(
-                        f"no base_image for VM '{new_vm_name}': "
-                        f"set base_image at the cluster or VM level"
-                    )
-                provider.clone_vm(
-                    src_vm_name=src_vm_name,
-                    new_vm_name=new_vm_name,
-                    info=vm_info,
-                    workdir=cluster['workdir']
-                )
-            return
-        except Exception:
-            if not last_attempt:
-                delay = attempt * 2
-                log.warning(
-                    f"clone {new_vm_name} failed (attempt {attempt}/{max_retries}), "
-                    f"retrying in {delay}s"
-                )
-                time.sleep(delay)
-            else:
-                raise
 
 
-class BoxmanManager(ConfigMixin, WorkspaceMixin, NamingMixin, SSHMixin):
+class BoxmanManager(ConfigMixin, WorkspaceMixin, NamingMixin, SSHMixin, ImagesMixin, NetworksMixin, VMsMixin):
     def __init__(self,
                  config: dict[str, Any] | None = None):
         """
@@ -659,891 +609,6 @@ class BoxmanManager(ConfigMixin, WorkspaceMixin, NamingMixin, SSHMixin):
         """
         return self.runtime_instance.inject_into_provider_config(provider_config)
 
-    def import_image(self, cli_args) -> None:
-        """
-        Import an image into the provider's storage.
-
-        :param manager: The instance of the BoxmanManager
-        :param cli_args: The parsed arguments from the cli
-        """
-        # Phase 1 (#49): stays on the default session — import-image is a
-        # single-provider flow (the type comes from the manifest / CLI).
-        self.provider.import_image(
-            manifest_uri=cli_args.manifest_uri,
-            vm_name=cli_args.vm_name,
-            vm_dir=cli_args.vm_dir,
-            manifest_local_path=getattr(cli_args, 'manifest_local_path', None),
-        )
-
-    def push_image(self, cli_args) -> None:
-        """
-        Push a qcow2 image (and optional metadata) to an OCI registry.
-
-        Provider-agnostic — delegates to the ``oras`` CLI. ``self`` is unused
-        but kept for signature consistency with the other CLI dispatchers.
-        """
-        from boxman.providers.libvirt.oci_push import push_oci_image
-        try:
-            push_oci_image(
-                image_ref=cli_args.image_ref,
-                qcow2_path=cli_args.qcow2,
-                metadata_path=getattr(cli_args, 'metadata', None),
-            )
-            print(f"successfully pushed image to {cli_args.image_ref}", flush=True)
-        except (ValueError, RuntimeError) as exc:
-            print(f"error pushing image: {exc}", flush=True)
-            raise SystemExit(1) from exc
-
-    def inspect_image(self, cli_args) -> None:
-        """
-        Inspect an OCI image reference: print its manifest summary and, when a
-        ``vmimage.json`` sidecar is present, its metadata.
-
-        Provider-agnostic — delegates to the ``oras`` CLI (manifest fetch, no
-        full blob download). ``self`` is unused but kept for signature
-        consistency with the other CLI dispatchers.
-        """
-        from boxman.providers.libvirt.oci_pull import (
-            format_inspect,
-            inspect_oci_image,
-        )
-        try:
-            summary = inspect_oci_image(cli_args.image_ref)
-            print(format_inspect(summary), end="", flush=True)
-        except (ValueError, RuntimeError) as exc:
-            print(f"error inspecting image: {exc}", flush=True)
-            raise SystemExit(1) from exc
-
-    def pxe_boot(self, cli_args):
-        """
-        Set boot order to network-first on a VM, start it, optionally wait
-        for SSH, and optionally restore the boot order afterwards.
-
-        Designed to be used with a Cobbler PXE provisioning server.
-        """
-        session = self.provider  # Phase 1 (#49): single-VM PXE flow stays on the default session until Phase 3
-        vm_name = cli_args.vm
-
-        self.logger.info(f"setting boot order to [network, hd] for '{vm_name}'")
-        if not session.set_boot_order(vm_name, ['network', 'hd']):
-            self.logger.error(f"failed to set boot order for '{vm_name}'")
-            return False
-
-        self.logger.info(f"starting VM '{vm_name}'")
-        if not session.start_vm(vm_name):
-            self.logger.error(f"failed to start VM '{vm_name}'")
-            return False
-
-        if cli_args.expected_ip:
-            ok = session.wait_for_ssh(
-                cli_args.expected_ip,
-                timeout=cli_args.wait_timeout,
-            )
-            if not ok:
-                self.logger.error(
-                    f"SSH timeout waiting for '{vm_name}' at "
-                    f"{cli_args.expected_ip}")
-                return False
-
-            if cli_args.restore_after:
-                self.logger.info(
-                    f"restoring boot order to [hd] for '{vm_name}'")
-                session.restore_boot_order(vm_name)
-
-        return True
-
-    def create_templates(self, cli_args) -> None:
-        """
-        Create template VMs from cloud images using cloud-init.
-
-        Reads the ``templates`` section from the project config and creates
-        each template VM that doesn't already exist.
-
-        :param cli_args: The parsed arguments from the cli
-        """
-        requested = None
-        if cli_args is not None and hasattr(cli_args, 'template_names') and cli_args.template_names:
-            requested = [t.strip() for t in cli_args.template_names.split(',')]
-
-        force = getattr(cli_args, 'force', False) if cli_args is not None else False
-
-        failed = self._create_templates_impl(requested=requested, force=force)
-        if failed:
-            self.logger.error(
-                f"{len(failed)} template(s) could not be created: "
-                f"{', '.join(failed)}")
-            raise SystemExit(1)
-
-    def _create_templates_impl(self, requested=None, force=False) -> list[str]:
-        """
-        Internal implementation for creating template VMs.
-
-        Returns:
-            The keys of the templates that could not be built. Empty when they
-            all succeeded. A caller that ignores this will happily clone from a
-            template whose cloud-init never finished.
-
-        Args:
-            requested: Optional list of template keys to create (None = all).
-            force: If True, recreate existing templates.
-        """
-        from boxman.image_cache import ImageCache
-        from boxman.providers.libvirt.cloudinit import CloudInitTemplate
-
-        cache_conf = self.app_config.get('cache', {}) if self.app_config else {}
-        image_cache = ImageCache.from_config(cache_conf)
-
-        config = self.config
-        templates = config.get('templates', {})
-
-        if not templates:
-            self.logger.warning("no templates defined in configuration")
-            return []
-
-        #: keys of the templates whose build failed, reported to the caller
-        failed: list[str] = []
-
-        # determine provider config — templates are libvirt-only, so the
-        # libvirt block is resolved explicitly (app-merged + runtime-injected)
-        provider_config = self._libvirt_provider_config()
-
-        self.logger.info(f"resolved provider config for templates: {provider_config}")
-
-        # resolve a workdir for template artifacts
-        # Templates must live in a stable location independent of any cluster
-        # workdir, so that destroying or cleaning a cluster does not delete
-        # the template disk images that other clusters may still reference.
-        default_workdir = '~/boxman-templates'
-
-        # --- Pre-check: detect already-existing templates ----------------
-        # Build a temporary VirshCommand to query existing VMs once.
-        _virsh = VirshCommand(provider_config=provider_config)
-        _existing_vms: set = set()
-        result = _virsh.execute("list", "--all", "--name", hide=True, warn=True)
-        if result.ok:
-            _existing_vms = {
-                v.strip() for v in result.stdout.strip().split("\n") if v.strip()
-            }
-
-        # Identify which of the requested templates already exist
-        existing_templates: list[str] = []
-        templates_to_create: list[str] = []
-
-        for tpl_key, tpl_conf in templates.items():
-            if requested and tpl_key not in requested:
-                continue
-            tpl_name = tpl_conf.get('name', tpl_key)
-            if tpl_name in _existing_vms:
-                existing_templates.append(tpl_key)
-            else:
-                templates_to_create.append(tpl_key)
-
-        # If any templates already exist and --force was NOT given, error out.
-        if existing_templates and not force:
-            names = ", ".join(
-                f"'{templates[k].get('name', k)}'" for k in existing_templates
-            )
-            self.logger.error(
-                f"the following template(s) already exist: {names}. "
-                f"Use --force to delete and recreate them."
-            )
-            return list(existing_templates)
-
-        # Merge both lists (existing ones will be force-recreated)
-        all_keys = existing_templates + templates_to_create
-        # -----------------------------------------------------------------
-
-        for tpl_key in all_keys:
-            tpl_conf = templates[tpl_key]
-
-            tpl_name = tpl_conf.get('name', tpl_key)
-            image_field = tpl_conf.get('image', '')
-            if isinstance(image_field, dict):
-                image_path = image_field.get('uri', '')
-                image_checksum = image_field.get('checksum', None)
-            else:
-                image_path = image_field
-                image_checksum = None
-
-            # Common typo: 'file' instead of 'image'
-            if not image_path and 'file' in tpl_conf:
-                image_path = tpl_conf['file']
-                self.logger.warning(
-                    f"template '{tpl_key}': 'file' is not a valid key, "
-                    f"did you mean 'image'? Using '{image_path}' as the image path.")
-
-            cloudinit_userdata = tpl_conf.get('cloudinit', None)
-            cloudinit_metadata = tpl_conf.get('cloudinit_metadata', None)
-            cloudinit_network_config = tpl_conf.get('cloudinit_network_config', None)
-            cloudinit_done_marker = tpl_conf.get('cloudinit_done_marker', None)
-            cloudinit_agent_timeout = tpl_conf.get('cloudinit_agent_timeout', 300)
-            cloudinit_guest_exec_timeout = tpl_conf.get('cloudinit_guest_exec_timeout', 120)
-            cloudinit_done_timeout = tpl_conf.get('cloudinit_done_timeout', 120)
-            cloudinit_fallback_timeout = tpl_conf.get('cloudinit_fallback_timeout', 180)
-            tpl_memory = tpl_conf.get('memory', 2048)
-            tpl_vcpus = tpl_conf.get('vcpus', 2)
-            tpl_os_variant = tpl_conf.get('os_variant', 'generic')
-            tpl_disk_format = tpl_conf.get('disk_format', 'qcow2')
-            tpl_disk_size = tpl_conf.get('disk_size', None)
-            tpl_network = tpl_conf.get('network', 'default')
-            tpl_bridge = tpl_conf.get('bridge', None)
-            tpl_workdir = tpl_conf.get('workdir', default_workdir)
-
-            # Ensure the workdir exists and is writable by the current user.
-            # Earlier steps (e.g. docker runtime) may have created it as root.
-            expanded_workdir = os.path.expanduser(tpl_workdir)
-            self._ensure_writable_dir(expanded_workdir)
-
-            # Also pre-create the template subdirectory that cloudinit.py
-            # will use, so it doesn't hit PermissionError.
-            template_subdir = os.path.join(expanded_workdir, tpl_name)
-            self._ensure_writable_dir(template_subdir)
-
-            self.logger.info(f"creating template '{tpl_key}' -> VM name '{tpl_name}'")
-
-            ct = CloudInitTemplate(
-                template_name=tpl_name,
-                image_path=image_path,
-                cloudinit_userdata=cloudinit_userdata,
-                cloudinit_metadata=cloudinit_metadata,
-                cloudinit_network_config=cloudinit_network_config,
-                cloudinit_done_marker=cloudinit_done_marker,
-                cloudinit_agent_timeout=cloudinit_agent_timeout,
-                cloudinit_guest_exec_timeout=cloudinit_guest_exec_timeout,
-                cloudinit_done_timeout=cloudinit_done_timeout,
-                cloudinit_fallback_timeout=cloudinit_fallback_timeout,
-                workdir=tpl_workdir,
-                provider_config=provider_config,
-                memory=tpl_memory,
-                vcpus=tpl_vcpus,
-                os_variant=tpl_os_variant,
-                disk_format=tpl_disk_format,
-                disk_size=tpl_disk_size,
-                network=tpl_network,
-                bridge=tpl_bridge,
-                image_checksum=image_checksum,
-                image_cache=image_cache,
-            )
-
-            try:
-                success = ct.create_template(force=force)
-            except ValueError as exc:
-                # a bad timeout or marker in the template block
-                self.logger.error(f"template '{tpl_key}': {exc}")
-                success = False
-
-            if success:
-                self.logger.info(f"template '{tpl_key}' created successfully")
-            else:
-                self.logger.error(f"failed to create template '{tpl_key}'")
-                failed.append(tpl_key)
-
-        return failed
-
-    def _ensure_writable_dir(self, path: str) -> None:
-        """
-        Ensure *path* exists and is writable by the current user.
-
-        If the directory was created by another user (e.g. root via docker),
-        attempt to fix ownership with ``sudo chown``.  If ``sudo`` is not
-        available or fails, a clear error message is logged.
-
-        When running under a non-local runtime the directory is also
-        created inside the container so that commands executed via
-        ``docker exec`` can access it.
-
-        Args:
-            path: Absolute or user-expandable directory path.
-        """
-        path = os.path.expanduser(path)
-
-        if not os.path.exists(path):
-            try:
-                os.makedirs(path, exist_ok=True)
-            except PermissionError:
-                # parent dir may be owned by root — try sudo mkdir
-                self.logger.warning(
-                    f"cannot create '{path}' as current user, "
-                    f"trying with sudo...")
-                result = run(
-                    f"sudo mkdir -p '{path}'", hide=True, warn=True)
-                if not result.ok:
-                    raise PermissionError(
-                        f"failed to create directory '{path}' even with sudo"
-                    ) from None
-                # fall through to chown below
-
-        # Two failure modes to repair here:
-        #   1. The directory itself is not writable by us (e.g. created
-        #      as root by docker compose mount).
-        #   2. The directory IS writable but contains stale entries
-        #      owned by another user (typically root, left over from a
-        #      previous docker-runtime run). Tools like `genisoimage`
-        #      open their output with O_TRUNC and need write access on
-        #      the existing file, not just the parent dir.
-        self._normalize_ownership(path)
-
-        # When using a non-local runtime (e.g. docker-compose), also
-        # create the directory inside the container so that commands
-        # executed via 'docker exec' can write to it.
-        if self._runtime_name != 'local':
-            mkdir_cmd = self.runtime_instance.wrap_command(
-                f"mkdir -p '{path}'"
-            )
-            self.logger.info(f"creating directory inside runtime container: {path}")
-            result = run(mkdir_cmd, hide=True, warn=True)
-            if not result.ok:
-                self.logger.warning(
-                    f"failed to create '{path}' inside container: "
-                    f"{result.stderr.strip()}")
-
-        # Mark the directory as owned by the current runtime so that a
-        # later cross-runtime reuse triggers the collision prompt.
-        self._write_runtime_sentinel(path, self._runtime_name)
-
-    def _normalize_ownership(self, path: str) -> None:
-        """
-        Make sure *path* and its top-level entries are usable by the
-        current user.
-
-        Strategy (cheapest path first, sudo last resort):
-
-        1. If the directory itself is not writable by us, escalate
-           straight to ``sudo chown -R``.
-        2. Else, scan the top-level entries; if any are owned by
-           another user, try to remove them (``unlink`` for files,
-           ``shutil.rmtree`` for dirs). This works WITHOUT sudo as
-           long as the parent dir is user-writable, because ``unlink``
-           only needs write+exec on the parent, not on the file.
-        3. If unlink/rmtree fails (e.g. nested foreign-owned tree we
-           can't traverse), fall back to ``sudo chown -R``.
-        4. If sudo also fails, raise ``PermissionError`` with a
-           copy-pasteable fix command.
-
-        Foreign-owned files inside boxman workdirs are always either
-        stale build artifacts (seed.iso, qcow2 disk images) or stale
-        provisioning files we are about to regenerate, so removing
-        them is safe.
-        """
-        my_uid = os.getuid()
-        my_gid = os.getgid()
-
-        # Fast path: directory itself isn't writable → straight to sudo.
-        dir_writable = os.access(path, os.W_OK)
-
-        foreign_entries: list = []
-        if dir_writable:
-            try:
-                for entry in os.scandir(path):
-                    try:
-                        if entry.stat(
-                            follow_symlinks=False
-                        ).st_uid != my_uid:
-                            foreign_entries.append(entry)
-                    except OSError:
-                        # Can't stat → treat as foreign so we attempt
-                        # the recovery path.
-                        foreign_entries.append(entry)
-            except OSError:
-                # Can't scandir → fall through to sudo chown.
-                dir_writable = False
-
-        if dir_writable and not foreign_entries:
-            return
-
-        # Try the cheap path first: just unlink/rmtree the foreign
-        # entries. No sudo needed when the parent dir is writable.
-        unrecoverable: list = []
-        if dir_writable and foreign_entries:
-            for entry in foreign_entries:
-                try:
-                    if entry.is_dir(follow_symlinks=False):
-                        shutil.rmtree(entry.path)
-                    else:
-                        os.unlink(entry.path)
-                    self.logger.info(
-                        f"removed stale foreign-owned entry: "
-                        f"{entry.path}")
-                except OSError as exc:
-                    unrecoverable.append((entry.path, exc))
-
-            if not unrecoverable:
-                return
-
-        # Fall back to sudo chown -R.
-        self.logger.info(
-            f"fixing ownership of '{path}' to {my_uid}:{my_gid} "
-            f"via sudo chown -R "
-            f"(directory not writable or contained foreign entries "
-            f"that could not be removed)"
-        )
-        result = run(
-            f"sudo -n chown -R {my_uid}:{my_gid} '{path}'",
-            hide=True, warn=True,
-        )
-        if result.ok:
-            return
-
-        # Sudo failed — emit an actionable error so the user knows
-        # exactly what to fix.
-        offenders = "\n  ".join(
-            sorted(set(
-                [path]
-                + [p for p, _ in unrecoverable]
-                + [e.path for e in foreign_entries]
-            ))
-        )
-        stderr = (result.stderr or "").strip()
-        raise PermissionError(
-            f"could not normalise ownership of '{path}'.\n"
-            f"sudo chown failed: {stderr or '(no stderr)'}\n"
-            f"\n"
-            f"Run one of these to fix manually:\n"
-            f"  sudo chown -R {my_uid}:{my_gid} '{path}'\n"
-            f"  sudo rm -rf '{path}'   "
-            f"# safe — boxman will recreate it\n"
-            f"\n"
-            f"Affected paths:\n  {offenders}"
-        )
-
-    @staticmethod
-    def _is_diskless_boot(vm_info: dict) -> bool:
-        """True if a VM boots from network (PXE) or cdrom (ISO).
-
-        Such VMs are created directly via ``virt-install`` and need no
-        ``base_image`` to clone from.
-        """
-        boot_order = vm_info.get('boot_order', ['hd'])
-        return bool(boot_order) and boot_order[0] in ('network', 'cdrom')
-
-    @staticmethod
-    def _validate_cdrom_boot(vm_info: dict, iso_names: set) -> tuple[bool, str]:
-        """Validate an ISO-boot VM's ``cdroms:`` up front.
-
-        Returns ``(ok, reason)``. An ISO-boot VM must declare a first cdrom that
-        either references a known ``isos:`` entry or carries an explicit local
-        ``source:``.
-        """
-        cdroms = vm_info.get('cdroms')
-        if not cdroms:
-            return False, "boot_order starts with 'cdrom' but no 'cdroms:' are defined"
-        first = cdroms[0]
-        if isinstance(first, str):
-            name = first
-        elif isinstance(first, dict) and first.get('name'):
-            name = first['name']
-        elif isinstance(first, dict) and first.get('source'):
-            return True, ""  # explicit local source, nothing to resolve
-        else:
-            return False, f"first cdroms entry {first!r} has neither a name nor a source"
-        if name not in iso_names:
-            return False, f"cdroms references unknown iso '{name}' (declare it under isos:)"
-        return True, ""
-
-    def validate_base_images(self) -> None:
-        """
-        Validate VM boot configuration up front, before any parallel cloning.
-
-        - ``hd``-boot VMs (the default) must have a ``base_image`` — from the VM
-          or inherited from the cluster.
-        - ``cdrom``-boot (ISO) VMs must declare a valid ``cdroms:`` entry
-          referencing a known ``isos:`` entry (or an explicit ``source:``).
-        - ``network``-boot (PXE) VMs need neither.
-
-        Raises ``ValueError`` aggregating every problem found.
-        """
-        iso_names = set((self.config.get('isos') or {}).keys())
-        missing = []
-        invalid = []
-        for cluster_name, cluster in self.config.get('clusters', {}).items():
-            cluster_base = cluster.get('base_image', '')
-            for vm_name, vm_info in cluster.get('vms', {}).items():
-                loc = f"{cluster_name}.vms.{vm_name}"
-                boot_order = vm_info.get('boot_order', ['hd'])
-                first_boot = boot_order[0] if boot_order else 'hd'
-                if first_boot == 'cdrom':
-                    ok, reason = self._validate_cdrom_boot(vm_info, iso_names)
-                    if not ok:
-                        invalid.append(f"{loc} ({reason})")
-                    continue
-                if first_boot == 'network':
-                    continue  # PXE boot needs no base_image
-                if not vm_info.get('base_image') and not cluster_base:
-                    missing.append(loc)
-        errors = []
-        if missing:
-            errors.append(
-                "the following VM(s) have no base_image (set it at the cluster "
-                f"or VM level): {', '.join(missing)}")
-        if invalid:
-            errors.append(
-                "the following ISO-boot VM(s) have an invalid 'cdroms:' "
-                f"configuration: {', '.join(invalid)}")
-        if errors:
-            raise ValueError("; ".join(errors))
-
-    @staticmethod
-    def _oci_template_name(image_ref: str) -> str:
-        """Derive a deterministic, libvirt-safe template VM name from an OCI ref.
-
-        e.g. ``oci://registry.example.com/boxman/ubuntu-24.04:latest`` ->
-        ``boxman-oci-ubuntu-24.04-latest-<8 hex>``. The hash keeps distinct refs
-        that share a tag/name from colliding.
-        """
-        from boxman.providers.libvirt.oci_pull import _strip_scheme
-        ref = _strip_scheme(image_ref)
-        digest = hashlib.sha256(ref.encode('utf-8')).hexdigest()[:8]
-        readable = ref.rsplit('/', 1)[-1]  # e.g. 'ubuntu-24.04:latest'
-        safe = "".join(c if (c.isalnum() or c in '._-') else '-' for c in readable)
-        safe = safe.strip('-.') or 'image'
-        return f"boxman-oci-{safe}-{digest}"
-
-    def _expand_oci_base_images(self) -> None:
-        """Expand ``base_image: oci://…`` references into implicit templates.
-
-        The clone path can only clone an existing libvirt VM by name, so a direct
-        ``oci://`` base image cannot be cloned as-is. For each unique OCI ref found
-        in a cluster- or VM-level ``base_image``, synthesize a ``templates`` entry
-        that pulls the qcow2 from the registry (via the ``oci://`` template-image
-        path) and rewrite the ``base_image`` to that template's name. The existing
-        :meth:`ensure_templates_exist` + clone pipeline then handles the rest.
-
-        Idempotent: repeated refs map to the same template; non-OCI base images
-        are left untouched. Templates synthesized here specify no explicit
-        cloud-init, so the template build applies boxman's DEFAULT cloud-init
-        (default user, networking) — the pulled image should therefore be a
-        cloud-init-enabled cloud image. Use an explicit ``templates`` entry with
-        ``image.uri: oci://…`` when you need custom cloud-init.
-        """
-        clusters = self.config.get('clusters', {})
-        if not clusters:
-            return
-
-        templates = self.config.setdefault('templates', {})
-        ref_to_name: dict[str, str] = {}
-
-        def _resolve(base_image):
-            if not isinstance(base_image, str) or not base_image.startswith('oci://'):
-                return base_image
-            tpl_name = ref_to_name.get(base_image)
-            if tpl_name is None:
-                tpl_name = self._oci_template_name(base_image)
-                ref_to_name[base_image] = tpl_name
-                if tpl_name not in templates:
-                    templates[tpl_name] = {
-                        'name': tpl_name,
-                        'image': {'uri': base_image},
-                    }
-                    self.logger.info(
-                        f"expanded base_image '{base_image}' -> "
-                        f"implicit template '{tpl_name}'")
-            return tpl_name
-
-        for cluster in clusters.values():
-            if 'base_image' in cluster:
-                cluster['base_image'] = _resolve(cluster.get('base_image'))
-            for vm_info in cluster.get('vms', {}).values():
-                if 'base_image' in vm_info:
-                    vm_info['base_image'] = _resolve(vm_info.get('base_image'))
-
-    def _download_iso(self, url: str, dst_path: str) -> bool:
-        """Download an ISO from a URL, trying wget then curl then urllib.
-
-        Thin wrapper over :func:`boxman.utils.http_download.download_url`
-        (which shell-quotes URL/destination and uses ``curl --fail`` so an
-        HTTP 4xx/5xx error page is not accepted as a valid ISO).
-        """
-        self.logger.info(f"downloading ISO {url} -> {dst_path}")
-        if download_url(url, dst_path):
-            return True
-        self.logger.error(f"failed to download ISO from {url}")
-        return False
-
-    @staticmethod
-    def _iso_cache_filename(name: str, uri: str) -> str:
-        """Collision-free cache filename for an ISO.
-
-        Factory ISOs frequently share a basename (e.g. ``metal-amd64.iso``),
-        which would collide in the basename-keyed cache; disambiguate with the
-        declared iso name plus a short hash of the URI.
-        """
-        base = os.path.basename(urlparse(uri).path)
-        ext = os.path.splitext(base)[1] or ".iso"
-        safe = "".join(
-            c if (c.isalnum() or c in "._-") else "-" for c in str(name)
-        ).strip("-.") or "iso"
-        digest = hashlib.sha256(uri.encode("utf-8")).hexdigest()[:8]
-        return f"{safe}-{digest}{ext}"
-
-    def _resolve_isos(self) -> dict[str, str]:
-        """Download and cache all ISOs declared in the ``isos:`` config section.
-
-        Returns a mapping of iso_name -> local_file_path.
-        """
-        isos_conf = self.config.get("isos", {}) if self.config else {}
-        if not isos_conf:
-            return {}
-        if not isinstance(isos_conf, dict):
-            raise ValueError(
-                "'isos:' must be a mapping of <name>: {uri: ..., checksum: ...}, "
-                f"got {type(isos_conf).__name__}")
-
-        # ISO boot needs the file visible to the in-container virt-install; the
-        # host cache dir is not bind-mounted under a containerized runtime. Fail
-        # fast with guidance instead of a confusing missing-file error later.
-        runtime_name = getattr(self, "_runtime_name", "local") or "local"
-        if runtime_name != "local":
-            raise RuntimeError(
-                f"ISO boot ('isos:') is not yet supported under the "
-                f"'{runtime_name}' runtime; use the local runtime "
-                f"(the downloaded ISO is not visible inside the libvirt container)."
-            )
-
-        cache_conf = (self.app_config or {}).get("cache", {})
-        cache = ImageCache.from_config(cache_conf)
-
-        resolved: dict[str, str] = {}
-        for name, iso_conf in isos_conf.items():
-            if not isinstance(iso_conf, dict):
-                raise ValueError(
-                    f"iso '{name}' must be a mapping with at least a 'uri' key"
-                )
-            uri = iso_conf.get("uri")
-            if not uri:
-                raise ValueError(f"iso '{name}' missing 'uri'")
-            checksum = iso_conf.get("checksum")
-            filename = self._iso_cache_filename(name, uri)
-
-            local_path = cache.ensure(uri, self._download_iso, filename=filename)
-            if local_path is None:
-                if not cache.enabled:
-                    # Caching disabled: download directly to a stable path so the
-                    # ISO is still available to virt-install (re-downloaded each
-                    # run). Mirrors the base-image direct-download fallback.
-                    local_path = cache.cache_path_for(uri, filename=filename)
-                    os.makedirs(os.path.dirname(local_path), exist_ok=True)
-                    if not self._download_iso(uri, local_path):
-                        raise RuntimeError(
-                            f"Failed to download ISO '{name}' from {uri}")
-                else:
-                    raise RuntimeError(
-                        f"Failed to download ISO '{name}' from {uri}")
-
-            if checksum and not ImageCache.verify_checksum(local_path, checksum):
-                # Evict the bad file so a later run re-downloads instead of
-                # re-failing forever against the poisoned cache entry.
-                try:
-                    os.remove(local_path)
-                except OSError:
-                    pass
-                raise RuntimeError(f"Checksum mismatch for ISO '{name}'")
-
-            resolved[name] = local_path
-
-        return resolved
-
-    def _inject_resolved_iso(
-        self, vm_info: dict, resolved_isos: dict[str, str]
-    ) -> dict:
-        """Resolve ``cdroms:`` name references and inject ``_resolved_iso_path``.
-
-        Returns a shallow copy of vm_info with:
-        - Each cdrom entry expanded to carry ``source: <local_path>``. Entries
-          may be a plain string (an iso name), a ``{name: <iso>}`` mapping, or a
-          ``{source: <local path>}`` mapping; anything else is a clear error.
-        - ``_resolved_iso_path`` set to ``cdroms[0]['source']`` when
-          ``boot_order[0] == 'cdrom'``
-        """
-        cdroms = vm_info.get("cdroms", [])
-        if not cdroms:
-            return {**vm_info}
-
-        def _resolve_name(iso_name: str) -> str:
-            if iso_name not in resolved_isos:
-                raise ValueError(
-                    f"cdroms references unknown iso '{iso_name}'. "
-                    f"Declare it in the 'isos:' section."
-                )
-            return resolved_isos[iso_name]
-
-        resolved_cdroms = []
-        for cdrom in cdroms:
-            if isinstance(cdrom, str):
-                resolved_cdroms.append(
-                    {"name": cdrom, "source": _resolve_name(cdrom)})
-            elif isinstance(cdrom, dict) and cdrom.get("name"):
-                resolved_cdroms.append(
-                    {**cdrom, "source": _resolve_name(cdrom["name"])})
-            elif isinstance(cdrom, dict) and cdrom.get("source"):
-                resolved_cdroms.append(cdrom)
-            else:
-                raise ValueError(
-                    f"invalid cdroms entry {cdrom!r}: expected a string iso "
-                    f"name, a {{name: <iso>}} mapping, or a {{source: <path>}} "
-                    f"mapping"
-                )
-
-        result = {**vm_info, "cdroms": resolved_cdroms}
-
-        boot_order = vm_info.get("boot_order", ["hd"])
-        if boot_order and boot_order[0] == "cdrom" and resolved_cdroms:
-            first_source = resolved_cdroms[0].get("source")
-            if first_source:
-                result["_resolved_iso_path"] = first_source
-
-        return result
-
-    def _resolved_network_names(self, cluster_name: str, vm_info: dict) -> list[str]:
-        """Fully-qualified libvirt names for a VM's ``networks:`` (first-NIC list).
-
-        Direct-boot VMs (ISO/PXE) attach networks at ``virt-install`` time, so
-        the raw cluster-network name must be namespaced exactly as cluster
-        networks are defined (``bprj__…__clstr__…__<name>``).
-        """
-        names = []
-        for net in vm_info.get("networks") or []:
-            if isinstance(net, dict) and net.get("name"):
-                names.append(self.full_network_name(
-                    project_config=self.config,
-                    cluster_name=cluster_name,
-                    network_name=net["name"],
-                ))
-        return names
-
-    def _resolve_iso_config(self) -> None:
-        """Resolve ISO/cdrom/network references for direct-boot VMs, in place.
-
-        Downloads+caches declared ``isos:``, expands each VM's ``cdroms:`` to
-        local sources, sets ``_resolved_iso_path`` for cdrom-boot VMs, and
-        namespaces ``networks:`` into ``_resolved_networks``. Mutating
-        ``self.config`` means both the clone subprocesses and the later
-        configure/start step observe the resolved values (otherwise the resolved
-        ISO source never reaches the CDROM-attach path). Idempotent.
-        """
-        clusters = self.config.get("clusters", {})
-        if not clusters:
-            return
-        resolved_isos = self._resolve_isos()
-        for cluster_name, cluster in clusters.items():
-            for vm_name, vm_info in cluster.get("vms", {}).items():
-                resolved = self._inject_resolved_iso(vm_info, resolved_isos)
-                if self._is_diskless_boot(resolved):
-                    resolved["_resolved_networks"] = self._resolved_network_names(
-                        cluster_name, resolved)
-                cluster["vms"][vm_name] = resolved
-
-    def ensure_templates_exist(self) -> bool:
-        """
-        Check if any cluster's base_image refers to a template defined in the
-        ``templates`` section of the config. If the template VM does not exist,
-        create it automatically.
-
-        Returns:
-            True if all required templates exist (or were created), False on failure.
-        """
-        templates = self.config.get('templates', {})
-        if not templates:
-            return True  # nothing to do
-
-        # build a mapping: template VM name -> template key
-        tpl_name_to_key: dict[str, str] = {}
-        for tpl_key, tpl_conf in templates.items():
-            vm_name = tpl_conf.get('name', tpl_key)
-            tpl_name_to_key[vm_name] = tpl_key
-            # also map by the key itself in case base_image uses the key
-            tpl_name_to_key[tpl_key] = tpl_key
-
-        # collect which templates are referenced by clusters or individual VMs
-        needed_template_keys: set = set()
-        for _cluster_name, cluster in self.config.get('clusters', {}).items():
-            base_image = cluster.get('base_image', '')
-            if base_image in tpl_name_to_key:
-                needed_template_keys.add(tpl_name_to_key[base_image])
-            for _vm_name, vm_info in cluster.get('vms', {}).items():
-                vm_base = vm_info.get('base_image', '')
-                if vm_base and vm_base in tpl_name_to_key:
-                    needed_template_keys.add(tpl_name_to_key[vm_base])
-
-        if not needed_template_keys:
-            return True  # no cluster uses a template as base_image
-
-        # check which of the needed templates already exist as VMs
-        missing_keys: list = []      # template VM not registered at all
-        broken_keys: list = []       # registered, but backing disk gone
-        for tpl_key in needed_template_keys:
-            tpl_conf = templates[tpl_key]
-            tpl_vm_name = tpl_conf.get('name', tpl_key)
-
-            # ask the provider if the VM exists
-            # Phase 1 (#49): template management stays on the default
-            # session — templates are project-level and libvirt-only.
-            exists = False
-            if self.provider is not None and hasattr(self.provider, 'vm_exists'):
-                exists = self.provider.vm_exists(tpl_vm_name)
-            elif self.provider is not None:
-                # fallback: try virsh list
-                try:
-                    result = self._virsh().execute("list", "--all", "--name", hide=True, warn=True)
-                    if result.ok:
-                        vm_list = [v.strip() for v in result.stdout.strip().split("\n") if v.strip()]
-                        exists = tpl_vm_name in vm_list
-                except Exception as exc:
-                    self.logger.warning(f"could not check if template VM '{tpl_vm_name}' exists: {exc}")
-
-            if exists:
-                # The domain is registered, but its backing qcow2 file may
-                # have been deleted out from under libvirt. virt-clone then
-                # fails inscrutably during provision; detect it up-front and
-                # rebuild via the existing force-recreate path which already
-                # does destroy + undefine --remove-all-storage on its own.
-                disks_ok = True
-                if (self.provider is not None
-                        and hasattr(self.provider, 'template_disks_present')):
-                    try:
-                        disks_ok = self.provider.template_disks_present(tpl_vm_name)
-                    except Exception as exc:
-                        self.logger.warning(
-                            f"could not verify template '{tpl_vm_name}' disks: {exc}")
-                if not disks_ok:
-                    self.logger.warning(
-                        f"template VM '{tpl_vm_name}' is registered with libvirt "
-                        f"but its disk file is missing on the host — rebuilding "
-                        f"the template (force=True will destroy the orphan domain)")
-                    broken_keys.append(tpl_key)
-                else:
-                    self.logger.info(
-                        f"template VM '{tpl_vm_name}' already exists, skipping creation")
-            else:
-                self.logger.info(
-                    f"template VM '{tpl_vm_name}' (key='{tpl_key}') does not exist, "
-                    f"will create it before provisioning")
-                missing_keys.append(tpl_key)
-
-        if not missing_keys and not broken_keys:
-            return True
-
-        failed: list[str] = []
-        if missing_keys:
-            self.logger.info(
-                f"auto-creating {len(missing_keys)} missing template(s): {missing_keys}")
-            failed += self._create_templates_impl(
-                requested=missing_keys, force=False)
-        if broken_keys:
-            self.logger.info(
-                f"auto-rebuilding {len(broken_keys)} broken template(s): {broken_keys}")
-            failed += self._create_templates_impl(
-                requested=broken_keys, force=True)
-
-        if failed:
-            # returning True here would let provisioning carry on and clone
-            # from a template whose cloud-init never finished
-            self.logger.error(
-                f"template(s) {', '.join(failed)} could not be built, so the "
-                f"VMs that use them cannot be cloned")
-            return False
-
-        return True
-
     ### register/un-register the project in the cache
     def register_project_in_cache(self) -> None:
         """
@@ -1842,798 +907,6 @@ class BoxmanManager(ConfigMixin, WorkspaceMixin, NamingMixin, SSHMixin):
         """``docker compose down --volumes`` every docker-compose cluster."""
         for cluster_name, cluster in self._compose_clusters.items():
             self.session_for_cluster(cluster_name).destroy_cluster(cluster_name, cluster)
-
-    def define_networks(self) -> None:
-        """
-        Define the networks specified in the cluster configuration (sequential).
-
-        Networks must be defined one at a time because bridge name assignment
-        (virbrX) is a shared resource: each definition must be committed to
-        libvirt and the cache before the next network picks its bridge name,
-        otherwise two concurrent processes can both select the same bridge index
-        and the second net-define fails with "bridge name already in use".
-        """
-        for cluster_name, cluster in self._vm_clusters.items():
-            for network_name, network_info in (cluster.get('networks') or {}).items():
-                _network_name = self.full_network_name(
-                    project_config=self.config,
-                    cluster_name=cluster_name,
-                    network_name=network_name
-                )
-                self.session_for_cluster(cluster_name).define_network(
-                    name=_network_name,
-                    info=network_info,
-                    workdir=cluster['workdir']
-                )
-                self.logger.status(f"defined network {_network_name} in {cluster['workdir']}")
-
-    def reconcile_networks(self,
-                           dry_run: bool = False,
-                           allow_recreate: bool = False,
-                           auto_accept: bool = False) -> dict[str, str]:
-        """
-        Bring the libvirt networks in line with the configuration.
-
-        Three outcomes per network, decided by comparing the configuration
-        against ``virsh net-dumpxml``:
-
-        - **create** -- the network is not defined yet. Defined and started.
-        - **live** -- only the dhcp reservations or the dhcp range differ.
-          Applied with ``virsh net-update ... --live --config``: dnsmasq is
-          reloaded, the bridge stays up, no guest is disturbed.
-        - **recreate** -- the forward mode, address, netmask, bridge or mac
-          differ. libvirt cannot change these in place, so the network has to
-          be destroyed and defined again. That deletes the bridge and leaves
-          attached guests with a dead nic, so it only happens with
-          *allow_recreate*, and the guests are re-attached afterwards.
-
-        Args:
-            dry_run: report the plan and change nothing
-            allow_recreate: permit the disruptive path
-            auto_accept: skip the confirmation prompt for a recreate
-
-        Returns:
-            A mapping of network name to the action taken: one of ``created``,
-            ``updated``, ``recreated``, ``skipped`` or ``failed``.
-        """
-        if not hasattr(self.provider, 'plan_network'):
-            self.logger.debug(
-                "provider does not support network reconciliation, skipping")
-            return {}
-
-        results: dict[str, str] = {}
-
-        for cluster_name, cluster in self.config['clusters'].items():
-            for network_name, network_info in (cluster.get('networks') or {}).items():
-                full_name = self.full_network_name(
-                    project_config=self.config,
-                    cluster_name=cluster_name,
-                    network_name=network_name)
-
-                try:
-                    plan = self.provider.plan_network(
-                        name=full_name, info=network_info)
-                except ValueError as exc:
-                    # the network block itself does not validate (a bad dhcp
-                    # reservation, say). Report it against the network it came
-                    # from and carry on: the other networks, and the VMs, are
-                    # not necessarily affected
-                    self.logger.error(
-                        f"network {network_name}: {exc}")
-                    results[full_name] = 'failed'
-                    continue
-
-                if plan['action'] == 'none':
-                    continue
-
-                # two clusters may each hold a network called 'mgmt', so the
-                # logs carry the cluster as well
-                label = f"{cluster_name}/{network_name}"
-
-                if plan['action'] == 'error':
-                    results[full_name] = 'failed'
-                    continue
-
-                for line in net_reconcile.describe_plan(label, plan):
-                    self.logger.info(line)
-
-                if plan['action'] == 'create':
-                    self.logger.info(f"network {label}: not defined yet")
-                    if dry_run:
-                        results[full_name] = 'skipped'
-                        continue
-                    results[full_name] = self._define_network(
-                        label=label, full_name=full_name,
-                        network_info=network_info, workdir=cluster['workdir'])
-
-                elif plan['action'] == 'live':
-                    if dry_run:
-                        if plan.get('inactive'):
-                            self.logger.info(
-                                f"[dry-run] network {label}: is defined but "
-                                f"not running, would start it")
-                        results[full_name] = 'skipped'
-                        continue
-
-                    ok = True
-                    if plan.get('inactive'):
-                        self.logger.info(
-                            f"network {label}: is defined but not running, "
-                            f"starting it")
-                        ok = self.provider.start_network(
-                            name=full_name, info=network_info)
-
-                    if ok and (plan['host_ops'] or plan['range_ops']):
-                        ok = self.provider.apply_network_live_plan(
-                            name=full_name, info=network_info, plan=plan)
-                    results[full_name] = 'updated' if ok else 'failed'
-
-                elif plan['action'] == 'recreate':
-                    results[full_name] = self._recreate_network(
-                        cluster=cluster,
-                        network_name=label,
-                        full_name=full_name,
-                        network_info=network_info,
-                        plan=plan,
-                        dry_run=dry_run,
-                        allow_recreate=allow_recreate,
-                        auto_accept=auto_accept)
-
-        return results
-
-    def report_network_results(self, results: dict[str, str]) -> None:
-        """
-        Log the outcome of a reconcile, loudly for the ones that went wrong.
-
-        ``failed`` and ``partial`` would otherwise be buried: the caller
-        carries on either way, so this is the only place a user learns that a
-        network did not come back or that a guest is still disconnected.
-        """
-        if not results:
-            return
-
-        for full_name, outcome in sorted(results.items()):
-            if outcome == 'failed':
-                self.logger.error(f"network {full_name}: {outcome}")
-            elif outcome == 'partial':
-                self.logger.warning(
-                    f"network {full_name}: recreated, but at least one VM "
-                    f"could not be reconnected")
-            else:
-                self.logger.info(f"network {full_name}: {outcome}")
-
-    def _vms_worth_waiting_for(self) -> list[str]:
-        """
-        Project VMs minus the ones a recreate could not reconnect.
-
-        Waiting on a guest we already reported as unreachable only burns the
-        whole timeout before saying what we already knew.
-        """
-        unreachable = getattr(self, '_reattach_failed_vms', set())
-        return sorted(set(self._get_project_vm_names()) - unreachable)
-
-    def wait_for_vm_ips(self, vm_names: list[str], max_wait: int = 300) -> bool:
-        """
-        Wait until every named VM reports an address, or *max_wait* passes.
-
-        Returns:
-            True if they all got one.
-        """
-        if not vm_names:
-            return True
-
-        self.logger.info("waiting for VMs to get IP addresses...")
-        wait_time = 1
-        total_waited = 0
-        while total_waited < max_wait:
-            if all(self.provider.get_vm_ip_addresses(name) for name in vm_names):
-                self.logger.info(
-                    f"all VMs have IP addresses (waited {total_waited}s)")
-                return True
-            time.sleep(wait_time)
-            total_waited += wait_time
-            wait_time = min(wait_time * 2, 60)
-
-        self.logger.warning(
-            f"not every VM had an IP address after {max_wait}s; the ssh "
-            f"config may be incomplete")
-        return False
-
-    def _define_network(self,
-                        label: str,
-                        full_name: str,
-                        network_info: dict[str, Any],
-                        workdir: str) -> str:
-        """
-        Define a network, turning a conflict into a result instead of a crash.
-
-        ``define_network`` starts with ``check_network_exists()``, which raises
-        when the cache already holds an entry with this name, bridge or
-        address. That is the right guard when two projects collide, but it
-        raises before the try inside ``define_network``, so left alone it
-        reaches the CLI as a traceback.
-        """
-        try:
-            ok = self.provider.define_network(
-                name=full_name, info=network_info, workdir=workdir)
-        except RuntimeError as exc:
-            self.logger.error(f"network {label}: could not be defined: {exc}")
-            return 'failed'
-        return 'created' if ok else 'failed'
-
-    def _forget_cached_network(self, full_name: str) -> None:
-        """
-        Drop a network's entry from the projects cache.
-
-        Needed before redefining it: ``check_network_exists()`` walks every
-        cached project *including this one*, so a network's own leftover entry
-        counts as a conflict with itself -- same name, same address -- and the
-        redefine raises instead of running.
-        """
-        try:
-            if self.cache.unregister_network(self.config['project'], full_name):
-                self.logger.debug(f"removed {full_name} from the projects cache")
-        except (KeyError, OSError, ValueError) as exc:
-            # not fatal on its own, but the redefine that follows will fail on
-            # the stale entry, so say why
-            self.logger.warning(
-                f"could not drop {full_name} from the projects cache: {exc}")
-
-    @staticmethod
-    def _teardown_info(plan: dict[str, Any],
-                       network_info: dict[str, Any]) -> dict[str, Any]:
-        """
-        Describe the network **as it is now**, for the removal step.
-
-        The iptables rules to withdraw are the ones that were installed for the
-        current definition, so they follow its forward mode, bridge and subnet
-        -- not the ones being defined in its place. Handing ``remove_network``
-        the new configuration would, on a ``nat`` -> ``route`` change, try to
-        withdraw route rules that were never added and leave the nat rules
-        behind.
-
-        No dhcp block is carried over: nothing in the teardown reads it, and
-        reservations validated against the new subnet would be rejected when
-        paired with the old address.
-        """
-        actual = plan.get('actual') or {}
-        if not actual.get('mode'):
-            return network_info
-
-        info: dict[str, Any] = {'mode': actual['mode']}
-
-        if actual.get('bridge_name'):
-            info['bridge'] = {'name': actual['bridge_name']}
-        if actual.get('ip_address'):
-            info['ip'] = {'address': actual['ip_address']}
-            if actual.get('netmask'):
-                info['ip']['netmask'] = actual['netmask']
-
-        return info
-
-    def _recreate_network(self,
-                          cluster: dict[str, Any],
-                          network_name: str,
-                          full_name: str,
-                          network_info: dict[str, Any],
-                          plan: dict[str, Any],
-                          dry_run: bool,
-                          allow_recreate: bool,
-                          auto_accept: bool) -> str:
-        """
-        Destroy and redefine one network, then reconnect its guests.
-
-        Split out of :meth:`reconcile_networks` because the disruptive path is
-        where all the caveats live and it deserves to be read on its own.
-        """
-        attached = plan.get('attached_vms', [])
-        attached_text = ', '.join(attached) if attached else 'none'
-
-        if not allow_recreate:
-            self.logger.warning(
-                f"network {network_name}: the changes above need the network "
-                f"to be destroyed and defined again, which libvirt cannot do "
-                f"in place. Re-run with --recreate-networks to apply them "
-                f"(attached VMs that would be restarted: {attached_text})")
-            return 'skipped'
-
-        if dry_run:
-            self.logger.info(
-                f"[dry-run] would recreate network {network_name} and "
-                f"reconnect: {attached_text}")
-            return 'skipped'
-
-        if not auto_accept:
-            print(f"\nNetwork '{network_name}' has to be destroyed and "
-                  f"redefined to apply:")
-            for change in plan['structural']:
-                print(f"  - {change}")
-            print("\nIts bridge will be deleted and recreated. These VMs lose "
-                  f"their network link and will be reconnected, by a reboot "
-                  f"if their machine type cannot hot-plug: {attached_text}\n")
-            try:
-                answer = input(
-                    f"Type '{network_name}' to proceed: ").strip()
-            except EOFError:
-                # nothing is attached to stdin (a cron run, a pipeline): treat
-                # that as a no rather than a traceback
-                print("No input available, aborted.")
-                return 'skipped'
-            if answer != network_name:
-                print("Aborted.")
-                return 'skipped'
-
-        self.logger.info(f"network {network_name}: removing")
-        removed = True
-        try:
-            removed = self.provider.remove_network(
-                name=full_name, info=self._teardown_info(plan, network_info))
-        except RuntimeError as exc:
-            # remove_network destroys and undefines before it touches iptables,
-            # so the network is already gone: say what was left behind rather
-            # than aborting half way
-            self.logger.warning(
-                f"network {network_name}: removed, but its firewall rules "
-                f"could not be cleaned up: {exc}")
-
-        if not removed:
-            # destroy or undefine failed, so the network is still there.
-            # Redefining on top of it would fail confusingly
-            self.logger.error(
-                f"network {network_name}: could not be removed, leaving it "
-                f"as it is rather than defining on top of it")
-            return 'failed'
-
-        # the cache still lists the network we just removed, and
-        # check_network_exists() would count that as a conflict with itself
-        self._forget_cached_network(full_name)
-
-        self.logger.info(f"network {network_name}: defining again")
-        if self._define_network(
-                label=network_name, full_name=full_name,
-                network_info=network_info,
-                workdir=cluster['workdir']) == 'failed':
-            self.logger.error(
-                f"network {network_name}: could not be defined again. The "
-                f"attached VMs are left disconnected: {attached_text}")
-            return 'failed'
-
-        reattach_failed = []
-        for domain in attached:
-            outcome = self.provider.reattach_domain_network(domain, full_name)
-            if outcome == 'failed':
-                reattach_failed.append(domain)
-                # remembered so the post-recreate IP wait does not sit on a
-                # guest we already know is not coming back on its own
-                if not hasattr(self, '_reattach_failed_vms'):
-                    self._reattach_failed_vms = set()
-                self._reattach_failed_vms.add(domain)
-                self.logger.error(
-                    f"{domain}: could not be reconnected to {network_name}, "
-                    f"start it by hand")
-
-        if reattach_failed:
-            # the network is back but not every guest is, and saying
-            # 'recreated' would paper over a VM that is still down
-            return 'partial'
-
-        return 'recreated'
-
-    def destroy_networks(self) -> dict:
-        """
-        Destroy the networks specified in the cluster configuration (parallel).
-
-        Returns:
-            The ``failures`` dict from :meth:`_run_parallel`, empty when
-            every network was removed.
-        """
-        def _destroy(cluster_name, cluster, network_name, network_info):
-            _network_name = self.full_network_name(
-                project_config=self.config,
-                cluster_name=cluster_name,
-                network_name=network_name
-            )
-            self.session_for_cluster(cluster_name).remove_network(
-                name=_network_name,
-                info=network_info
-            )
-            self.logger.info(f"removed network {_network_name} in {cluster['workdir']}")
-            xml_path = os.path.expanduser(
-                os.path.join(cluster['workdir'], f'{_network_name}_net_define.xml'))
-            if os.path.isfile(xml_path):
-                os.remove(xml_path)
-                self.logger.info(f"removed network XML {xml_path}")
-
-        processes = [
-            (f"{cluster_name}/{network_name}", _destroy,
-             (cluster_name, cluster, network_name, network_info))
-            for cluster_name, cluster in self._vm_clusters.items()
-            for network_name, network_info in (cluster.get('networks') or {}).items()
-        ]
-        _results, failures = self._run_parallel(
-            processes, op_label='destroy network')
-        return failures
-    ### end networks define / remove / destroy
-
-    ### vms define / remove / destroy
-    def _ensure_libvirt_storage_pool(self, workdir: str, cluster_name: str) -> None:
-        """
-        Ensure the libvirt directory storage pool exists for the given workdir.
-
-        virt-clone automatically tries to define a pool for the target directory
-        when given a --file path. When multiple VMs are cloned in parallel into
-        the same directory, every process races to define the same pool and all
-        but the first fail with "pool already exists". Pre-defining the pool
-        here (sequentially, before parallel cloning begins) eliminates the race.
-
-        Uses *cluster_name*'s own libvirt session for the virsh connection so a
-        compose-primary mixed project (where the default ``self.provider`` is
-        the docker-compose session) still targets the configured libvirt
-        URI/runtime instead of silently falling back to local qemu:///system.
-        """
-        workdir = os.path.abspath(os.path.expanduser(workdir))
-        pool_name = os.path.basename(workdir)
-        virsh = VirshCommand(self.session_for_cluster(cluster_name).provider_config)
-
-        result = virsh.execute("pool-info", pool_name, warn=True)
-        if result.ok:
-            self.logger.info(f"storage pool '{pool_name}' already exists")
-            return
-
-        self.logger.info(f"defining storage pool '{pool_name}' for {workdir}")
-        virsh.execute("pool-define-as", pool_name, "dir", "--target", workdir)
-        virsh.execute("pool-build", pool_name, warn=True)
-        virsh.execute("pool-start", pool_name, warn=True)
-
-    def clone_vms(self) -> None:
-        """
-        Clone the VMs defined in the configuration.
-
-        The following is done for every vm in every cluster
-
-            - remove the vm
-            - clone the vm
-        """
-        def vm_clone_tasks():
-            prj_name = f'bprj__{self.config["project"]}__bprj'
-            for cluster_name, cluster in self._vm_clusters.items():
-                for vm_name, vm_info in cluster['vms'].items():
-                    vm_info = vm_info.copy()
-                    new_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-                    yield cluster, vm_info, new_vm_name
-
-        # Pre-define storage pools for all cluster workdirs before parallel
-        # cloning. virt-clone tries to auto-define a pool for the target
-        # directory; doing it here sequentially prevents the race condition
-        # where N parallel processes all try to define the same pool at once.
-        seen_workdirs: set = set()
-        for cluster_name, cluster in self._vm_clusters.items():
-            workdir = os.path.abspath(os.path.expanduser(cluster['workdir']))
-            if workdir not in seen_workdirs:
-                seen_workdirs.add(workdir)
-                self._ensure_libvirt_storage_pool(workdir, cluster_name)
-
-        # resolve isos/cdroms/networks in place so both the clone subprocesses
-        # and the later configure/start step see the resolved values
-        self._resolve_iso_config()
-        clone_tasks = [
-            (cluster, vm_info, new_vm_name)
-            for cluster, vm_info, new_vm_name in vm_clone_tasks()
-        ]
-        processes = [
-            Process(target=_clone_with_retry,
-                    args=(self.provider, cluster, vm_info, new_vm_name))
-            for cluster, vm_info, new_vm_name in clone_tasks
-        ]
-        [p.start() for p in processes]
-        [p.join() for p in processes]
-
-        # Abort provision if any clone subprocess exited non-zero. Without
-        # this check, the subsequent configure / start / wait-for-IP steps
-        # all spam errors against VMs that were never defined, and the
-        # wait-for-IP loop in particular looks like a hang.
-        failed = [
-            (task[2], p.exitcode)
-            for task, p in zip(clone_tasks, processes)
-            if p.exitcode != 0
-        ]
-        if failed:
-            names = ', '.join(name for name, _ in failed)
-            raise RuntimeError(
-                f"clone failed for {len(failed)} VM(s) ({names}); aborting "
-                f"provision. See the virt-clone error logged above for the "
-                f"underlying cause (typically a missing template disk or a "
-                f"libvirt storage pool issue).")
-
-    ### end vms define / remove / destroy
-
-    def _configure_and_start_vm(
-        self, cluster_name: str, cluster: dict[str, Any], vm_name: str, vm_info: dict[str, Any]
-    ) -> None:
-        """
-        Configure and start a single VM: cpu/mem, network interfaces, disks, then start.
-
-        Designed to be called in a separate process per VM after cloning is done.
-        """
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        full_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-        vm_info = vm_info.copy()
-
-        # cpu / memory
-        cpus = vm_info.get('cpus')
-        memory = vm_info.get('memory')
-        max_vcpus = vm_info.get('max_vcpus')
-        max_memory = vm_info.get('max_memory')
-        if cpus or memory:
-            self.logger.info(f"configuring cpu and memory for vm {vm_name}")
-            success = self.session_for_cluster(cluster_name).configure_vm_cpu_memory(
-                vm_name=full_vm_name, cpus=cpus, memory_mb=memory,
-                max_vcpus=max_vcpus, max_memory_mb=max_memory
-            )
-            if success:
-                self.logger.info(f"successfully configured cpu and memory for vm {vm_name}")
-            else:
-                self.logger.warning(f"failed to configure cpu and memory for vm {vm_name}")
-        else:
-            self.logger.warning(f"no cpu or memory configuration for vm {vm_name}, skipping")
-
-        # network interfaces
-        if 'network_adapters' not in vm_info:
-            self.logger.warning(f"no network adapters defined for vm {vm_name}, skipping")
-        else:
-            for adapter in vm_info['network_adapters']:
-                self.resolve_adapter_network(adapter, cluster_name)
-
-            self.logger.info(f"configuring network interfaces for vm {vm_name}")
-            success = self.session_for_cluster(cluster_name).configure_vm_network_interfaces(
-                vm_name=full_vm_name,
-                network_adapters=vm_info['network_adapters']
-            )
-            if success:
-                self.logger.info(f"network interfaces configured for vm {vm_name}")
-            else:
-                self.logger.warning(f"some network interfaces could not be configured for vm {vm_name}")
-
-        # disks
-        workdir = cluster.get('workdir', '.')
-        if 'disks' not in vm_info or not vm_info['disks']:
-            self.logger.warning(f"no disks defined for vm {vm_name}, skipping")
-        else:
-            self.logger.info(f"configuring disks for vm {vm_name}")
-            success = self.session_for_cluster(cluster_name).configure_vm_disks(
-                vm_name=full_vm_name,
-                disks=vm_info['disks'],
-                workdir=workdir,
-                disk_prefix=full_vm_name
-            )
-            if success:
-                self.logger.info(f"all disks configured for vm {vm_name}")
-            else:
-                self.logger.warning(f"some disks could not be configured for vm {vm_name}")
-
-        # shared folders (must be before start for virtiofs memfd backing)
-        if vm_info.get('shared_folders'):
-            self.logger.info(f"configuring shared folders for vm {vm_name}")
-            success = self.session_for_cluster(cluster_name).configure_vm_shared_folders(
-                vm_name=full_vm_name,
-                shared_folders=vm_info['shared_folders']
-            )
-            if success:
-                self.logger.info(f"shared folders configured for vm {vm_name}")
-            else:
-                self.logger.warning(f"some shared folders could not be configured for vm {vm_name}")
-
-        # cdroms — the cdrom-boot ISO is already attached by virt-install at
-        # create time, so attach only any *additional* cdroms here (avoids a
-        # spurious "missing source" failure and a duplicate attach)
-        boot_iso = vm_info.get('_resolved_iso_path')
-        extra_cdroms = [
-            c for c in (vm_info.get('cdroms') or [])
-            if not (isinstance(c, dict) and boot_iso and c.get('source') == boot_iso)
-        ]
-        if extra_cdroms:
-            self.logger.info(f"configuring CDROMs for vm {vm_name}")
-            success = self.session_for_cluster(cluster_name).configure_vm_cdroms(
-                vm_name=full_vm_name,
-                cdroms=extra_cdroms
-            )
-            if success:
-                self.logger.info(f"CDROMs configured for vm {vm_name}")
-            else:
-                self.logger.warning(f"some CDROMs could not be configured for vm {vm_name}")
-
-        # start
-        self.logger.info(f"starting vm {full_vm_name}")
-        success = self.session_for_cluster(cluster_name).start_vm(full_vm_name)
-        if success:
-            self.logger.info(f"successfully started the vm {full_vm_name}")
-        else:
-            self.logger.warning(f"failed to start the vm {full_vm_name}")
-
-    def _destroy_vm_and_disks(
-        self, cluster_name: str, cluster: dict[str, Any], vm_name: str, vm_info: dict[str, Any]
-    ) -> None:
-        """
-        Fully destroy a single VM: stop/undefine, remove disks, force-cleanup.
-
-        Designed to be called in a separate process per VM during deprovision.
-        """
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        full_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-        vm_info = vm_info.copy()
-
-        self.logger.info(f"destroying vm {full_vm_name}")
-        session = self.session_for_cluster(cluster_name)
-        session.destroy_vm(full_vm_name)
-        session.destroy_disks(
-            cluster['workdir'],
-            vm_name=full_vm_name,
-            disks=vm_info.get('disks', [])
-        )
-        session.destroy_vm(full_vm_name, force=True)
-
-    def _vm_disk_dirs(self, full_vm_name: str) -> list[str]:
-        """
-        Return the directories that hold *full_vm_name*'s disk files,
-        discovered from libvirt itself (``virsh domblklist``).
-
-        Used for VMs that no longer appear in conf.yml, whose workdir
-        can therefore not be resolved from the config. Falls back to
-        every workdir known to the project config when the domain
-        cannot be queried (e.g. already undefined) — the destroy_disks
-        glob is anchored at the VM name, so sweeping extra directories
-        is harmless.
-        """
-        virsh = VirshCommand(
-            provider_config=self.provider.provider_config)
-        result = virsh.execute(
-            "domblklist", full_vm_name, "--details", warn=True)
-        dirs = set()
-        if result.ok:
-            for row in parse_domblklist(result.stdout):
-                if row.device == 'disk' and row.source not in (None, '-'):
-                    dirs.add(os.path.dirname(row.source))
-        if not dirs:
-            self.logger.warning(
-                f"could not query disk paths for {full_vm_name} from "
-                f"libvirt; falling back to all configured workdirs")
-            dirs.update(self.collect_workdirs())
-        return sorted(dirs)
-
-    def _destroy_removed_vm(self, full_vm_name: str) -> None:
-        """
-        Destroy a VM that has been removed from the config.
-
-        Uses destroy_vm + destroy_disks with an empty disk list since we
-        no longer have the disk config for this VM. The disk directories
-        are read from libvirt before the domain is undefined (see
-        ``_vm_disk_dirs``); the glob-based cleanup in destroy_disks
-        catches all {vm_name}* artifacts.
-        """
-        self.logger.info(f"removing VM {full_vm_name} (no longer in config)")
-        # Phase 1 (#49): stays on the default session — the VM is gone
-        # from the config, so its cluster (and provider) can no longer be
-        # resolved. Revisited in Phase 3 (#51).
-        disk_dirs = self._vm_disk_dirs(full_vm_name)
-        self.provider.destroy_vm(full_vm_name)
-        for workdir in disk_dirs:
-            self.provider.destroy_disks(
-                workdir,
-                vm_name=full_vm_name,
-                disks=[]
-            )
-        self.provider.destroy_vm(full_vm_name, force=True)
-
-    def configure_and_start_vms(self) -> None:
-        """
-        Configure (cpu/mem, network interfaces, disks) and start all VMs in parallel.
-
-        Each VM is handled in its own process so all VMs are configured and
-        started concurrently.
-        """
-        processes = [
-            (f"{cluster_name}/{vm_name}", self._configure_and_start_vm,
-             (cluster_name, cluster, vm_name, vm_info))
-            for cluster_name, cluster in self._vm_clusters.items()
-            for vm_name, vm_info in cluster['vms'].items()
-        ]
-        self._run_parallel(processes, op_label='configure and start vm')
-
-    def _get_project_vm_names(self) -> list[str]:
-        """
-        Return the list of fully-qualified VM names that would be
-        created by provisioning the current config.
-        """
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        vm_names = []
-        for cluster_name, cluster in self.config.get('clusters', {}).items():
-            for vm_name in cluster.get('vms', {}).keys():
-                vm_names.append(f"{prj_name}_{cluster_name}_{vm_name}")
-        return vm_names
-
-    def _find_existing_project_vms(self) -> list[str]:
-        """
-        Query libvirt and return the subset of project VM names that
-        already exist (in any state).
-        """
-        expected = self._get_project_vm_names()
-        if not expected:
-            return []
-        if not self._has_libvirt_clusters():
-            return []
-
-        result = self._virsh().execute("list", "--all", "--name", hide=True, warn=True)
-        if not result.ok:
-            self.logger.warning("could not query existing VMs via virsh")
-            return []
-
-        existing = {
-            v.strip() for v in result.stdout.strip().split("\n") if v.strip()
-        }
-        return [vm for vm in expected if vm in existing]
-
-    def _find_all_existing_project_vms(self) -> list[str]:
-        """
-        Query libvirt and return ALL VM names that belong to this project
-        (match the project prefix), regardless of whether they appear in
-        the current config.
-        """
-        project = self.config.get("project", "")
-        prj_prefix = f"bprj__{project}__bprj_"
-
-        if not self._has_libvirt_clusters():
-            return []
-
-        result = self._virsh().execute("list", "--all", "--name", hide=True, warn=True)
-        if not result.ok:
-            self.logger.warning("could not query existing VMs via virsh")
-            return []
-
-        return [
-            v.strip() for v in result.stdout.strip().split("\n")
-            if v.strip() and v.strip().startswith(prj_prefix)
-        ]
-
-    def _get_vm_states(self) -> dict[str, str]:
-        """
-        Query libvirt and return a mapping of project VM name -> state string
-        for all project VMs that exist.
-
-        State strings are as returned by ``virsh list --all``, e.g.
-        'running', 'shut off', 'paused', 'saved', etc.
-
-        Returns:
-            Dict mapping full VM name to its state, only for VMs that exist.
-        """
-        expected = set(self._get_project_vm_names())
-        if not expected:
-            return {}
-        if not self._has_libvirt_clusters():
-            return {}
-
-        # Use the table output to get both name and state
-        result = self._virsh().execute("list", "--all", hide=True, warn=True)
-        if not result.ok:
-            self.logger.warning("could not query VM states via virsh")
-            return {}
-
-        states: dict[str, str] = {}
-        for line in result.stdout.strip().splitlines():
-            # Skip header and separator lines
-            line = line.strip()
-            if not line or line.startswith('---') or line.startswith('Id'):
-                continue
-            # Format: " Id   Name                 State"
-            # e.g.  " -    my-vm                shut off"
-            #       " 3    my-vm                running"
-            parts = line.split(None, 2)
-            if len(parts) >= 3:
-                vm_name = parts[1]
-                vm_state = parts[2].strip()
-                if vm_name in expected:
-                    states[vm_name] = vm_state
-
-        return states
 
     def provision(self, cli_args):
 
@@ -4630,473 +2903,3 @@ class BoxmanManager(ConfigMixin, WorkspaceMixin, NamingMixin, SSHMixin):
             return
         user = getattr(cli_args, "user", None)
         print(self.netlab.ssh_command(node_name, user=user))
-
-    ### end netlab CLI handlers ####
-
-    ### update (runtime modification) functions ####
-
-    def _clone_and_configure_new_vms(self, new_vm_names: set) -> None:
-        """
-        Clone and configure only the VMs whose full names are in new_vm_names.
-
-        Reuses the same logic as provision (clone + configure + start) but
-        filters to just the specified VMs.
-        """
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-
-        # pre-define storage pools for workdirs of new VMs
-        seen_workdirs: set = set()
-        for cluster_name, cluster in self._vm_clusters.items():
-            for vm_name in cluster['vms']:
-                full = f"{prj_name}_{cluster_name}_{vm_name}"
-                if full in new_vm_names:
-                    workdir = os.path.abspath(os.path.expanduser(cluster['workdir']))
-                    if workdir not in seen_workdirs:
-                        seen_workdirs.add(workdir)
-                        self._ensure_libvirt_storage_pool(workdir, cluster_name)
-
-        # clone new VMs (parallel with retry)
-        # resolve isos/cdroms/networks in place so both the clone subprocesses
-        # and the configure/start step below see the resolved values
-        self._resolve_iso_config()
-        clone_tasks = []
-        for cluster_name, cluster in self._vm_clusters.items():
-            for vm_name, vm_info in cluster['vms'].items():
-                full = f"{prj_name}_{cluster_name}_{vm_name}"
-                if full in new_vm_names:
-                    clone_tasks.append((cluster, vm_info, full))
-
-        processes = [
-            Process(target=_clone_with_retry,
-                    args=(self.provider, cluster, vm_info, new_vm_name))
-            for cluster, vm_info, new_vm_name in clone_tasks
-        ]
-        [p.start() for p in processes]
-        [p.join() for p in processes]
-
-        # Abort the update if any clone subprocess exited non-zero — same
-        # guard as :meth:`clone_vms` to prevent downstream configure/start
-        # steps from running against VMs that were never defined.
-        failed = [
-            (task[2], p.exitcode)
-            for task, p in zip(clone_tasks, processes)
-            if p.exitcode != 0
-        ]
-        if failed:
-            names = ', '.join(name for name, _ in failed)
-            raise RuntimeError(
-                f"clone failed for {len(failed)} new VM(s) ({names}); "
-                f"aborting update. See the virt-clone error logged above.")
-
-        # configure and start new VMs (parallel)
-        configure_tasks = []
-        for cluster_name, cluster in self._vm_clusters.items():
-            for vm_name, vm_info in cluster['vms'].items():
-                full = f"{prj_name}_{cluster_name}_{vm_name}"
-                if full in new_vm_names:
-                    configure_tasks.append((cluster_name, cluster, vm_name, vm_info))
-
-        processes = [
-            Process(
-                target=self._configure_and_start_vm,
-                args=(cluster_name, cluster, vm_name, vm_info)
-            )
-            for cluster_name, cluster, vm_name, vm_info in configure_tasks
-        ]
-        [p.start() for p in processes]
-        [p.join() for p in processes]
-
-    def _update_single_vm(
-        self,
-        cluster_name: str,
-        cluster: dict[str, Any],
-        vm_name: str,
-        vm_info: dict[str, Any],
-        result_queue: Queue,
-        dry_run: bool = False,
-    ) -> None:
-        """
-        Diff and apply updates to a single existing VM. Runs in its own process.
-
-        Results are put into result_queue as:
-            (vm_name, {'status': 'no_change'|'updated'|'needs_restart'|'failed', 'details': str})
-        """
-        from boxman.providers.libvirt.vm_differ import VMStateDiffer
-
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        full_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-        workdir = os.path.abspath(os.path.expanduser(cluster.get('workdir', '.')))
-
-        try:
-            differ = VMStateDiffer(provider_config=self.provider.provider_config)
-            diff = differ.diff_vm(
-                domain_name=full_vm_name,
-                desired_cpus=vm_info.get('cpus'),
-                desired_memory_mb=vm_info.get('memory'),
-                desired_disks=vm_info.get('disks'),
-                workdir=workdir,
-                disk_prefix=full_vm_name,
-                desired_max_vcpus=vm_info.get('max_vcpus'),
-                desired_max_memory_mb=vm_info.get('max_memory'),
-                desired_shared_folders=vm_info.get('shared_folders'),
-                desired_cdroms=vm_info.get('cdroms'),
-            )
-
-            has_changes = (
-                diff['cpu_changed'] or
-                diff['memory_changed'] or
-                diff['max_vcpus_changed'] or
-                diff['max_memory_changed'] or
-                diff['new_disks'] or
-                diff['resize_disks'] or
-                diff['new_cdroms'] or
-                diff['removed_cdroms'] or
-                diff['changed_cdroms'] or
-                diff['new_shared_folders'] or
-                diff['removed_shared_folders'] or
-                diff['changed_shared_folders']
-            )
-
-            if not has_changes:
-                self.logger.info(f"VM {vm_name}: no changes detected")
-                result_queue.put((vm_name, {'status': 'no_change', 'details': ''}))
-                return
-
-            # log the diff
-            changes = []
-            if diff['cpu_changed']:
-                changes.append(
-                    f"CPU: {diff['actual_cpus']} -> {diff['desired_cpus']}")
-            if diff['memory_changed']:
-                changes.append(
-                    f"memory: {diff['actual_memory_mb']}M -> {diff['desired_memory_mb']}M")
-            if diff['max_vcpus_changed']:
-                changes.append(
-                    f"max_vcpus: {diff['actual_max_vcpus']} -> {diff['desired_max_vcpus']}")
-            if diff['max_memory_changed']:
-                changes.append(
-                    f"max_memory: {diff['actual_max_memory_mb']}M -> {diff['desired_max_memory_mb']}M")
-            if diff['new_disks']:
-                names = [d.get('name', '?') for d in diff['new_disks']]
-                changes.append(f"new disks: {', '.join(names)}")
-            if diff['resize_disks']:
-                resizes = [
-                    f"{r['target']} {r['current_size_mb']}M->{r['desired_size_mb']}M"
-                    for r in diff['resize_disks']
-                ]
-                changes.append(f"resize disks: {', '.join(resizes)}")
-            if diff['new_cdroms']:
-                names = [c.get('name', '?') for c in diff['new_cdroms']]
-                changes.append(f"new cdroms: {', '.join(names)}")
-            if diff['removed_cdroms']:
-                targets = [c['target'] for c in diff['removed_cdroms']]
-                changes.append(f"remove cdroms: {', '.join(targets)}")
-            if diff['changed_cdroms']:
-                swaps = [f"{c['target']}->{c['source']}" for c in diff['changed_cdroms']]
-                changes.append(f"change cdroms: {', '.join(swaps)}")
-            if diff['new_shared_folders']:
-                names = [f.get('name', '?') for f in diff['new_shared_folders']]
-                changes.append(f"new shared folders: {', '.join(names)}")
-            if diff['removed_shared_folders']:
-                names = [f['name'] for f in diff['removed_shared_folders']]
-                changes.append(f"remove shared folders: {', '.join(names)}")
-            if diff['changed_shared_folders']:
-                names = [f.get('name', '?') for f in diff['changed_shared_folders']]
-                changes.append(f"change shared folders: {', '.join(names)}")
-
-            self.logger.info(f"VM {vm_name}: changes detected: {'; '.join(changes)}")
-
-            if dry_run:
-                result_queue.put((vm_name, {
-                    'status': 'dry_run',
-                    'details': '; '.join(changes)
-                }))
-                return
-
-            # apply changes
-            vm_running = diff['vm_state'] == 'running'
-            restart_needed = False
-
-            # CPU / memory / max ceilings
-            if (diff['cpu_changed'] or diff['memory_changed'] or
-                    diff['max_vcpus_changed'] or diff['max_memory_changed']):
-                cpu_mem_result = self.provider.update_vm_cpu_memory(
-                    vm_name=full_vm_name,
-                    cpus=diff['desired_cpus'] if diff['cpu_changed'] else None,
-                    memory_mb=diff['desired_memory_mb'] if diff['memory_changed'] else None,
-                    vm_state=diff['vm_state'],
-                    actual_cpus=diff['actual_cpus'],
-                    actual_memory_mb=diff['actual_memory_mb'],
-                    max_vcpus=diff.get('desired_max_vcpus'),
-                    max_memory_mb=diff.get('desired_max_memory_mb')
-                )
-                if not cpu_mem_result['success']:
-                    result_queue.put((vm_name, {
-                        'status': 'failed',
-                        'details': 'CPU/memory update failed'
-                    }))
-                    return
-                if cpu_mem_result['restart_needed']:
-                    restart_needed = True
-
-            # disks
-            if diff['new_disks'] or diff['resize_disks']:
-                disk_ok = self.provider.update_vm_disks(
-                    vm_name=full_vm_name,
-                    new_disks=diff['new_disks'],
-                    resize_disks=diff['resize_disks'],
-                    workdir=workdir,
-                    disk_prefix=full_vm_name,
-                    vm_running=vm_running
-                )
-                if not disk_ok:
-                    result_queue.put((vm_name, {
-                        'status': 'failed',
-                        'details': 'disk update failed'
-                    }))
-                    return
-
-            # cdroms
-            if diff['new_cdroms'] or diff['removed_cdroms'] or diff['changed_cdroms']:
-                cdrom_ok = self.provider.update_vm_cdroms(
-                    vm_name=full_vm_name,
-                    new_cdroms=diff['new_cdroms'],
-                    removed_cdroms=diff['removed_cdroms'],
-                    changed_cdroms=diff['changed_cdroms'],
-                    vm_running=vm_running
-                )
-                if not cdrom_ok:
-                    result_queue.put((vm_name, {
-                        'status': 'failed',
-                        'details': 'CDROM update failed'
-                    }))
-                    return
-
-            # shared folders
-            if (diff['new_shared_folders'] or diff['removed_shared_folders'] or
-                    diff['changed_shared_folders']):
-                folder_result = self.provider.update_vm_shared_folders(
-                    vm_name=full_vm_name,
-                    new_folders=diff['new_shared_folders'],
-                    removed_folders=diff['removed_shared_folders'],
-                    changed_folders=diff['changed_shared_folders'],
-                    vm_running=vm_running
-                )
-                if not folder_result['success']:
-                    result_queue.put((vm_name, {
-                        'status': 'failed',
-                        'details': 'shared folder update failed'
-                    }))
-                    return
-                if folder_result.get('restart_needed'):
-                    restart_needed = True
-
-            # handle restart if needed
-            if restart_needed and vm_running:
-                self.logger.info(
-                    f"VM {vm_name}: restarting to apply changes "
-                    f"(live max ceiling cannot be raised)")
-                self.provider.shutdown_and_wait(full_vm_name)
-                self.provider.start_vm(full_vm_name)
-                result_queue.put((vm_name, {
-                    'status': 'updated',
-                    'details': '; '.join(changes) + ' (restarted)'
-                }))
-            else:
-                result_queue.put((vm_name, {
-                    'status': 'updated',
-                    'details': '; '.join(changes)
-                }))
-
-        except Exception as exc:
-            self.logger.error(f"VM {vm_name}: update failed: {exc}")
-            result_queue.put((vm_name, {
-                'status': 'failed',
-                'details': str(exc)
-            }))
-
-    def update(self, cli_args):
-        """
-        Apply config changes to already-provisioned VMs.
-
-        Compares the desired state in conf.yml against actual VM state in
-        libvirt and applies only the changes needed:
-          - New VMs in config are cloned, configured, and started
-          - CPU/memory changes are applied (hot if possible, cold otherwise)
-          - New disks are created and attached
-          - Existing disks are resized (grow only)
-          - Removed VMs (in libvirt but no longer in config) are destroyed
-
-        Use --dry-run to preview changes without applying them.
-        Use --yes to skip the confirmation prompt for VM removal.
-        """
-        config = self.config
-        dry_run = getattr(cli_args, 'dry_run', False)
-        auto_accept = getattr(cli_args, 'yes', False)
-
-        # ensure provider configs reflect runtime settings
-        # Phase 1 (#49): the update/diff flow below stays on the default
-        # session — it is deeply libvirt-shaped (VMStateDiffer, virsh
-        # edits) and only libvirt clusters can exist until Phase 3 (#51).
-        self._update_sessions_with_runtime()
-
-        # --- networks first ---
-        # a new VM further down may be wired to a network that does not exist
-        # yet, and this runs before the early return below so that a change
-        # which only touches networks is not silently a no-op
-        self.ensure_shared_bridges()
-        network_results = self.reconcile_networks(
-            dry_run=dry_run,
-            allow_recreate=getattr(cli_args, 'recreate_networks', False),
-            auto_accept=auto_accept)
-
-        self.report_network_results(network_results)
-
-        # --- categorize VMs ---
-        expected_vms = set(self._get_project_vm_names())
-        all_existing_vms = set(self._find_all_existing_project_vms())
-
-        new_vm_names = expected_vms - all_existing_vms
-        update_vm_names = expected_vms & all_existing_vms
-        removed_vm_names = all_existing_vms - expected_vms
-
-        if not new_vm_names and not update_vm_names and not removed_vm_names:
-            self.logger.info("no VMs to add, update, or remove")
-            return
-
-        # --- summary ---
-        if new_vm_names:
-            short = [n.split('_')[-1] for n in sorted(new_vm_names)]
-            self.logger.info(f"VM(s) to add: {', '.join(short)}")
-        if update_vm_names:
-            short = [n.split('_')[-1] for n in sorted(update_vm_names)]
-            self.logger.info(f"VM(s) to update: {', '.join(short)}")
-        if removed_vm_names:
-            short = [n.split('_')[-1] for n in sorted(removed_vm_names)]
-            self.logger.info(f"VM(s) to remove: {', '.join(short)}")
-
-        # --- confirmation for destructive removal ---
-        if removed_vm_names and not dry_run and not auto_accept:
-            short = [n.split('_')[-1] for n in sorted(removed_vm_names)]
-            print(
-                f"\nThe following VM(s) will be permanently destroyed: "
-                f"{', '.join(short)}")
-            print(
-                "This will stop the VM(s), remove their disks, and "
-                "clean up all associated resources.\n")
-            answer = input("Proceed? [y/N] ").strip().lower()
-            if answer not in ("y", "yes"):
-                print("Aborted.")
-                return
-
-        # --- handle new VMs ---
-        if new_vm_names:
-            if dry_run:
-                self.logger.info("[dry-run] would clone and configure new VMs")
-            else:
-                # expand any `base_image: oci://…` into implicit templates
-                # before resolving/cloning (the clone path needs a VM name).
-                self._expand_oci_base_images()
-                # ensure templates exist -- cloning from one that failed to
-                # build produces VMs whose cloud-init never ran
-                if not self.ensure_templates_exist():
-                    self.logger.error(
-                        "aborting: not every template could be created")
-                    return
-                try:
-                    self.validate_base_images()
-                except ValueError as exc:
-                    self.logger.error(str(exc))
-                    return
-
-                self._clone_and_configure_new_vms(new_vm_names)
-
-                # wait for IPs on new VMs
-                self.logger.info("waiting for new VMs to get IP addresses...")
-                self.wait_for_vm_ips(new_vm_names, max_wait=300)
-
-                # eject cdrom on new VMs
-                for vm_name in new_vm_names:
-                    self.provider.eject_cdrom(vm_name)
-
-        # --- handle existing VMs ---
-        if update_vm_names:
-            prj_name = f'bprj__{config["project"]}__bprj'
-            result_queue: Queue = Queue()
-
-            update_tasks = []
-            for cluster_name, cluster_cfg in self._vm_clusters.items():
-                for vm_name, vm_info in cluster_cfg['vms'].items():
-                    full = f"{prj_name}_{cluster_name}_{vm_name}"
-                    if full in update_vm_names:
-                        update_tasks.append(
-                            (cluster_name, cluster_cfg, vm_name, vm_info))
-
-            processes = [
-                Process(
-                    target=self._update_single_vm,
-                    args=(cluster_name, cluster_cfg, vm_name, vm_info,
-                          result_queue, dry_run)
-                )
-                for cluster_name, cluster_cfg, vm_name, vm_info in update_tasks
-            ]
-            [p.start() for p in processes]
-            [p.join() for p in processes]
-
-            # collect and print results
-            results = {}
-            while not result_queue.empty():
-                vm_name, result = result_queue.get()
-                results[vm_name] = result
-
-            # print summary
-            no_change = [n for n, r in results.items() if r['status'] == 'no_change']
-            updated = [n for n, r in results.items() if r['status'] == 'updated']
-            failed = [n for n, r in results.items() if r['status'] == 'failed']
-            dry_run_items = [n for n, r in results.items() if r['status'] == 'dry_run']
-
-            if dry_run_items:
-                self.logger.info("--- dry-run summary ---")
-                for vm_name in dry_run_items:
-                    self.logger.info(f"  {vm_name}: {results[vm_name]['details']}")
-
-            if no_change:
-                self.logger.info(f"no changes: {', '.join(no_change)}")
-            if updated:
-                self.logger.info(f"updated: {', '.join(updated)}")
-                for vm_name in updated:
-                    self.logger.info(f"  {vm_name}: {results[vm_name]['details']}")
-            if failed:
-                self.logger.error(f"failed: {', '.join(failed)}")
-                for vm_name in failed:
-                    self.logger.error(f"  {vm_name}: {results[vm_name]['details']}")
-
-        # --- handle removed VMs ---
-        if removed_vm_names:
-            if dry_run:
-                self.logger.info("[dry-run] would destroy the following VMs:")
-                for vm_name in sorted(removed_vm_names):
-                    self.logger.info(f"  {vm_name}")
-            else:
-                processes = [
-                    (vm_name, self._destroy_removed_vm, (vm_name,))
-                    for vm_name in removed_vm_names
-                ]
-                self._run_parallel(processes, op_label='destroy removed vm')
-
-                short = [n.split('_')[-1] for n in sorted(removed_vm_names)]
-                self.logger.info(
-                    f"removed {len(removed_vm_names)} VM(s): "
-                    f"{', '.join(short)}")
-
-        # regenerate SSH config and display connect info
-        if not dry_run:
-            # a recreated network power-cycles the guests attached to it, so
-            # the addresses the ssh config is written from do not exist yet
-            if any(outcome in ('recreated', 'partial')
-                   for outcome in network_results.values()):
-                self.wait_for_vm_ips(self._vms_worth_waiting_for())
-            self.setup_ssh_access()
-            self.connect_info()
-
-    ### end update functions ####
